+++ src/DiscoToolbarBundle.php
@@ -14,31 +14,23 @@ declare(strict_types=1);
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
  * @copyright 2025 Marcin Orlowski
  * @license   https://opensource.org/license/mit MIT
- * @link      https://github.com/MarcinOrlowski/php-symfony-discodevbar
+ * @link      https://github.com/MarcinOrlowski/php-discotoolbar-symfony
  *
  * ########################################################################## */
 
-namespace MarcinOrlowski\DiscoDevBar\DependencyInjection;
+namespace MarcinOrlowski\DiscoToolbar;
 
-use Symfony\Component\Config\FileLocator;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Extension\Extension;
-use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
- * DiscoDevBar Extension
+ * DiscoToolbar Bundle
  *
- * Loads bundle services configuration
+ * Provides customizable toolbar for Symfony projects
  */
-class DiscoDevBarExtension extends Extension
+class DiscoToolbarBundle extends Bundle
 {
-    public function load(array $configs, ContainerBuilder $container): void
+    public function getPath(): string
     {
-        $loader = new YamlFileLoader(
-            $container,
-            new FileLocator(\dirname(__DIR__) . '/../Resources/config')
-        );
-
-        $loader->load('services.yaml');
+        return \dirname(__DIR__);
     }
 }